Move parser and scanner into their own files.
